// YOUCOM.ReserVook/src/app/features/master/codename/model/codename.model.ts
import { Base } from '../../../../shared/model/baseinfo.model'

export class CodeNameInfo implements Base {
  companyNo: string;
  divisionCode: string;
  code: string;
  codeName: string;
  codeValue: string;
  displayOrder: number;
  status: string;
  version :number;
  creator:string;
  updator:string;
  cdt:string;
  udt:string;
}